requirements.txt
ipykernel
ipywidgets
torch
transformers
trl
scikit-learn
matplotlib
evaluate
peft
wandb
load_dotenv
python-box
pymilvus
sentence-transformers
langchain
langchain_community
langchain_milvus
unsloth
triton==2.3.0
absl-py 
nltk 
rouge-score
sacrebleu
openai
